browser/base/content/test/test_bug395533.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=395533
-->
<head>
  <title>Test for Bug 395533</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=395533">Mozilla Bug 395533</a>
<p id="display"><iframe id="testFrame" src="bug395533-data.txt"></iframe></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 395533 **/
SimpleTest.waitForExplicitFinish();

addLoadEvent(function() {
  // Need privs because the feed seems to have an about:feeds principal or some
  // such.  It's not same-origin with us in any case.
  netscape.security.PrivilegeManager.enablePrivilege("UniversalBrowserRead");
  is($("testFrame").contentDocument.documentElement.id, "",
     "Text got sniffed as a feed?");
});
addLoadEvent(SimpleTest.finish);




</script>
</pre>
</body>
</html>
